requirements.txt
Django==1.11.14
MySQL-python==1.2.5
django-mysql==1.1.1
Pillow==2.5.1
django-rest-swagger==2.1.0
djangorestframework==3.8.2
django-oauth-toolkit==0.10.0
requests==2.12.1
django-mptt==0.8.6
django-celery==3.2.1
pycrypto==2.6.1
pypinyin==0.8.5
# git+http://58.23.5.118:7990/projects/VND/repos/django-mobile/browse
